Done debugging to lab02 Exercise 4
1. Exercise 3 changes in function is_in_range(), seeing that
we need to add single quotation '' to digits when doing logical
compare with ASCII representation, or else digits will be recognize
to ASCII term. For example when you want to present zero as a
character, '0' should be the thing. No '' infront and back of 0
will be the null terminator of ASCII representation(\000).

2. One more modification is still in is_in_range, this time we
emphasize cgdb tool method for debugging. I will write down the
command myself without the help of lab description to memorize
the useful commands listed.

break, -b: break with [line number | function name]
conditional break, -b if <true | false>: break if the condition is satisfied
run, -r: execute the program until termination or reaching a breakpoint
continue, -c: continue the execution of a program that was paused
backtrace, -bt: print the road we step through to the place current at